fix(cli): derive the key census from the schema, keep it off the public surface
Review feedback on #5438. Two things were wrong with the first pass.

The known-key list was a second hand-maintained source sitting beside the
schema, so the next field added to one and not the other would work fine and
still warn that it was unsupported. It is now read off
`integrationSchema.shape`, so there is only one list.

`KNOWN_INTEGRATION_KEYS` and `unknownIntegrationKeys` were also reachable from
the published `./integration` subpath, which turns bookkeeping into API I can't
remove later. The schema and the census moved to
`authoring/integration/schema.mjs` — nothing in the exports map resolves there —
and `parse.mjs` exports `parseIntegration` and nothing else.

Also fixed the integration guide and the README, both of which still promised a
strict schema, which is the opposite of what this ships.
